requirements_optimized.txt
# Core Dependencies
anthropic>=0.40.0
openai>=1.50.0
groq>=0.11.0
google-generativeai>=0.8.0

# Performance & Async
aiohttp>=3.10.0
asyncio-throttle>=1.0.2
tenacity>=8.5.0
redis>=5.1.0
diskcache>=5.6.3

# Code Analysis & Security
bandit>=1.7.6
safety>=3.2.7
vulture>=2.11
mypy>=1.11.2
black>=24.8.0
isort>=5.13.2
flake8>=7.1.1
pylint>=3.2.7

# Development Tools
rich>=13.8.0
typer>=0.12.5
pydantic>=2.9.0
pydantic-settings>=2.5.2
streamlit>=1.39.0

# Git & Version Control
GitPython>=3.1.43
dulwich>=0.22.1

# Testing & Documentation  
pytest>=8.3.3
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
pdoc>=14.7.0
sphinx>=7.4.7

# Utilities
python-dotenv>=1.0.1
pathlib2>=2.3.7
click>=8.1.7
tqdm>=4.66.5
colorama>=0.4.6
tabulate>=0.9.0
